feat(earn): add page objects for EarnClaim, EarnDeposit, and EarnWithdraw features for streamlined test navigation and actions
